backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3

# Astronomy
lightkurve>=2.4
astropy>=6.0
astroquery>=0.4.7
numpy>=1.26,<2.0
scipy>=1.13
PyWavelets>=1.6

# ML
torch>=2.0
scikit-learn>=1.4

# Utilities
httpx>=0.27
pandas>=2.2
